Fix fetching waterfall after retrying a NeXus vis
