+++ src/h5web/vis-packs/nexus/hooks.ts
@@ -1,101 +1,8 @@
-import { useContext } from 'react';
-import type {
-  Dataset,
-  Group,
-  NumArrayDataset,
-  ScalarShape,
-  StringType,
-} from '../../providers/models';
-import { ProviderContext } from '../../providers/context';
-import {
-  isDefined,
-  assertArray,
-  assertDefined,
-  assertDataset,
-  assertNumericType,
-  assertArrayShape,
-  assertStringType,
-  assertScalarShape,
-} from '../../guards';
-import { getChildEntity } from '../../utils';
+import { isDefined } from '../../guards';
 import type { AxisMapping, ScaleType } from '../core/models';
 import { useDatasetValues } from '../core/hooks';
-import type { AxisDatasetMapping, NxData } from './models';
-import {
-  assertNxDataGroup,
-  findSignalDataset,
-  getDatasetLabel,
-  getAttributeValue,
-  getSilxStyle,
-  findErrorsDataset,
-} from './utils';
-
-function useTitleDataset(
-  group: Group
-): Dataset<ScalarShape, StringType> | undefined {
-  const { valuesStore } = useContext(ProviderContext);
-
-  const dataset = getChildEntity(group, 'title');
-  if (!dataset) {
-    return undefined;
-  }
-
-  assertDataset(dataset);
-  assertScalarShape(dataset);
-  assertStringType(dataset);
-
-  valuesStore.prefetch({ path: dataset.path });
-  return dataset;
-}
-
-function useAssociatedDatasets(
-  group: Group,
-  type: 'axes' | 'auxiliary_signals'
-): (NumArrayDataset | undefined)[] {
-  const { valuesStore } = useContext(ProviderContext);
-
-  const dsetList = getAttributeValue(group, type) || [];
-  const dsetNames = typeof dsetList === 'string' ? [dsetList] : dsetList;
-  assertArray<string>(dsetNames);
-
-  return dsetNames.map((name) => {
-    if (name === '.') {
-      return undefined;
-    }
-
-    const dataset = getChildEntity(group, name);
-    assertDefined(dataset);
-    assertDataset(dataset);
-    assertArrayShape(dataset);
-    assertNumericType(dataset);
-
-    valuesStore.prefetch({ path: dataset.path });
-    return dataset;
-  });
-}
-
-export function useNxData(group: Group): NxData {
-  assertNxDataGroup(group);
-  const signalDataset = findSignalDataset(group);
-  const errorsDataset = findErrorsDataset(group, signalDataset.name);
-
-  const { valuesStore } = useContext(ProviderContext);
-  valuesStore.prefetch({ path: signalDataset.path });
-  if (errorsDataset) {
-    valuesStore.prefetch({ path: errorsDataset.path });
-  }
-
-  return {
-    signalDataset,
-    errorsDataset,
-    titleDataset: useTitleDataset(group),
-    axisDatasets: useAssociatedDatasets(group, 'axes'),
-    silxStyle: getSilxStyle(group),
-    auxDatasets: useAssociatedDatasets(group, 'auxiliary_signals').filter(
-      isDefined
-    ),
-  };
-}
+import type { AxisDatasetMapping } from './models';
+import { getDatasetLabel } from './utils';
 
 export function useAxisMapping(
   mapping: AxisDatasetMapping,
