+++ src/h5web/vis-packs/nexus/utils.ts
@@ -6,6 +6,8 @@ import type {
   NumericType,
   ArrayShape,
   ComplexType,
+  ScalarShape,
+  StringType,
 } from '../../providers/models';
 import {
   assertDefined,
@@ -14,8 +16,12 @@ import {
   assertNumericType,
   assertArrayShape,
   assertNumericOrComplexType,
+  assertScalarShape,
+  assertStringType,
+  assertArray,
+  isDefined,
 } from '../../guards';
-import type { NxAttribute, SilxStyle } from './models';
+import type { NxAttribute, NxData, SilxStyle } from './models';
 import { isScaleType } from '../core/utils';
 import { getChildEntity } from '../../utils';
 
@@ -24,6 +30,20 @@ export function getAttributeValue(
   attributeName: NxAttribute
 ): unknown {
   return entity.attributes?.find((attr) => attr.name === attributeName)?.value;
+}
+
+export function getDatasetLabel(dataset: Dataset): string {
+  const longName = getAttributeValue(dataset, 'long_name');
+  if (longName && typeof longName === 'string') {
+    return longName;
+  }
+
+  const units = getAttributeValue(dataset, 'units');
+  if (units && typeof units === 'string') {
+    return `${dataset.name} (${units})`;
+  }
+
+  return dataset.name;
 }
 
 export function isNxDataGroup(group: Group) {
@@ -52,7 +72,7 @@ export function findSignalDataset(
   return dataset;
 }
 
-export function findErrorsDataset(
+function findErrorsDataset(
   group: Group,
   signalName: string
 ): NumArrayDataset | undefined {
@@ -70,18 +90,42 @@ export function findErrorsDataset(
   return dataset;
 }
 
-export function getDatasetLabel(dataset: Dataset): string {
-  const longName = getAttributeValue(dataset, 'long_name');
-  if (longName && typeof longName === 'string') {
-    return longName;
+function findAssociatedDatasets(
+  group: Group,
+  type: 'axes' | 'auxiliary_signals'
+): (NumArrayDataset | undefined)[] {
+  const dsetList = getAttributeValue(group, type) || [];
+  const dsetNames = typeof dsetList === 'string' ? [dsetList] : dsetList;
+  assertArray<string>(dsetNames);
+
+  return dsetNames.map((name) => {
+    if (name === '.') {
+      return undefined;
+    }
+
+    const dataset = getChildEntity(group, name);
+    assertDefined(dataset);
+    assertDataset(dataset);
+    assertArrayShape(dataset);
+    assertNumericType(dataset);
+
+    return dataset;
+  });
+}
+
+function findTitleDataset(
+  group: Group
+): Dataset<ScalarShape, StringType> | undefined {
+  const dataset = getChildEntity(group, 'title');
+  if (!dataset) {
+    return undefined;
   }
 
-  const units = getAttributeValue(dataset, 'units');
-  if (units && typeof units === 'string') {
-    return `${dataset.name} (${units})`;
-  }
+  assertDataset(dataset);
+  assertScalarShape(dataset);
+  assertStringType(dataset);
 
-  return dataset.name;
+  return dataset;
 }
 
 export function getSilxStyle(group: Group): SilxStyle {
@@ -111,4 +155,20 @@ export function getSilxStyle(group: Group): SilxStyle {
     console.warn(`Malformed 'SILX_style' attribute: ${silxStyle}`); // eslint-disable-line no-console
     return {};
   }
+}
+
+export function getNxData(group: Group): NxData {
+  assertNxDataGroup(group);
+  const signalDataset = findSignalDataset(group);
+  const errorsDataset = findErrorsDataset(group, signalDataset.name);
+  const auxDatasets = findAssociatedDatasets(group, 'auxiliary_signals');
+
+  return {
+    signalDataset,
+    errorsDataset,
+    axisDatasets: findAssociatedDatasets(group, 'axes'),
+    titleDataset: findTitleDataset(group),
+    silxStyle: getSilxStyle(group),
+    auxDatasets: auxDatasets.filter(isDefined),
+  };
 }
